+++ pisarz-clicker-firmware/src/clickEvent.h
@@ -19,7 +19,7 @@ typedef struct{
 
 } ClickEvent;
 
-ClickEvent clickEvent_create(uint16_t key, uint8_t *returnCode);
+ClickEvent clickEvent_create(uint8_t key, uint8_t *returnCode);
 ClickEvent clickEvent_createEmpty();
 void clickEvent_update(ClickEvent *event, ClickEvent *prevEvent);
